pyproject.toml
[tool.poetry]
name = "cryptowallet"
version = "0.1.0"
description = ""
authors = ["thelonggoodbuy <thelonggoodbuy1991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
mypy = "^1.8.0"
fastapi = "^0.109.0"
uvicorn = {extras = ["standart"], version = "^0.26.0"}
sqlalchemy = "^2.0.25"
sqlalchemy-file = "^0.6.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
sqlalchemy-utils = "^0.41.1"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
email-validator = "^2.1.0.post1"
password-validator = "^1.0"
fasteners = "^0.19"
aiofiles = "^23.2.1"
pydantic = "^2.6.1"
python-socketio = "^5.11.1"
redis = "^5.0.1"
aioredis = "2.0.0"
propan = {extras = ["async-rabbit"], version = "^0.1.5.26"}
aiokafka = "^0.10.0"
aio-pika = "^9.4.0"
web3 = "^6.15.1"
celery = "^5.3.6"
asgiref = "^3.8.1"
asyncpg = "^0.29.0"
aioetherscan = "^0.8.12"
pytz = "^2024.1"
aiohttp = {extras = ["websockets"], version = "^3.9.5"}
httpx = "^0.27.0"
websocket-client = "^1.8.0"
pre-commit = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
